+++ lib/load.php
@@ -18,6 +18,10 @@ if ( class_exists( 'WP_REST_Controller' ) ) {
 	if ( ! class_exists( 'WP_REST_Widget_Updater_Controller' ) ) {
 		require dirname( __FILE__ ) . '/class-wp-rest-widget-updater-controller.php';
 	}
+	if ( ! class_exists( 'WP_REST_Widget_Areas_Controller' ) ) {
+		require dirname( __FILE__ ) . '/class-experimental-wp-widget-blocks-manager.php';
+		require dirname( __FILE__ ) . '/class-wp-rest-widget-areas-controller.php';
+	}
 	/**
 	* End: Include for phase 2
 	*/