+++ packages/edit-post/src/index.native.js
@@ -1,4 +1,9 @@
 /**
+ * External dependencies
+ */
+import { Platform } from 'react-native';
+
+/**
  * WordPress dependencies
  */
 import '@wordpress/core-data';
@@ -18,12 +23,15 @@ export function initializeEditor() {
 	// register and setup blocks
 	registerCoreBlocks();
 
-	// disable Code and More blocks for the release
+	// disable Code block for the release
 	// eslint-disable-next-line no-undef
 	if ( typeof __DEV__ === 'undefined' || ! __DEV__ ) {
 		unregisterBlockType( 'core/code' );
-		unregisterBlockType( 'core/more' );
-		unregisterBlockType( 'core/video' );
+
+		// Disable Video block except for iOS for now.
+		if ( Platform.OS !== 'ios' ) {
+			unregisterBlockType( 'core/video' );
+		}
 	}
 }