+++ packages/edit-widgets/src/components/layout/index.js
@@ -9,30 +9,24 @@ import { navigateRegions } from '@wordpress/components';
  */
 import Header from '../header';
 import Sidebar from '../sidebar';
-import WidgetArea from '../widget-area';
+import WidgetAreas from '../widget-areas';
+import Notices from '../notices';
 
-function Layout() {
-	const areas = [
-		__( 'Sidebar' ),
-		__( 'Footer' ),
-		__( 'Header' ),
-	];
-
+function Layout( { blockEditorSettings } ) {
 	return (
 		<>
 			<Header />
 			<Sidebar />
+			<Notices />
 			<div
 				className="edit-widgets-layout__content"
 				role="region"
 				aria-label={ __( 'Widgets screen content' ) }
 				tabIndex="-1"
 			>
-				{ areas.map( ( area, index ) => (
-					<div key={ index } className="edit-widgets-layout__area">
-						<WidgetArea title={ area } initialOpen={ index === 0 } />
-					</div>
-				) ) }
+				<WidgetAreas
+					blockEditorSettings={ blockEditorSettings }
+				/>
 			</div>
 		</>
 	);