+++ packages/editor/src/components/provider/index.js
@@ -11,7 +11,7 @@ import { compose } from '@wordpress/compose';
 import { Component } from '@wordpress/element';
 import { withDispatch, withSelect } from '@wordpress/data';
 import { __ } from '@wordpress/i18n';
-import { BlockEditorProvider } from '@wordpress/block-editor';
+import { BlockEditorProvider, __experimentalTransformStyles } from '@wordpress/block-editor';
 import apiFetch from '@wordpress/api-fetch';
 import { addQueryArgs } from '@wordpress/url';
 import { decodeEntities } from '@wordpress/html-entities';
@@ -19,9 +19,10 @@ import { decodeEntities } from '@wordpress/html-entities';
 /**
  * Internal dependencies
  */
-import transformStyles from '../../editor-styles';
+import withRegistryProvider from './with-registry-provider';
 import { mediaUpload } from '../../utils';
 import ReusableBlocksButtons from '../reusable-blocks-buttons';
+import ConvertToGroupButtons from '../convert-to-group-buttons';
 
 const fetchLinkSuggestions = async ( search ) => {
 	const posts = await apiFetch( {
@@ -78,6 +79,7 @@ class EditorProvider extends Component {
 				'allowedBlockTypes',
 				'availableLegacyWidgets',
 				'bodyPlaceholder',
+				'codeEditingEnabled',
 				'colors',
 				'disableCustomColors',
 				'disableCustomFontSizes',
@@ -89,6 +91,7 @@ class EditorProvider extends Component {
 				'isRTL',
 				'maxWidth',
 				'styles',
+				'template',
 				'templateLock',
 				'titlePlaceholder',
 			] ),
@@ -109,7 +112,7 @@ class EditorProvider extends Component {
 			return;
 		}
 
-		const updatedStyles = transformStyles( this.props.settings.styles, '.editor-styles-wrapper' );
+		const updatedStyles = __experimentalTransformStyles( this.props.settings.styles, '.editor-styles-wrapper' );
 
 		map( updatedStyles, ( updatedCSS ) => {
 			if ( updatedCSS ) {
@@ -158,12 +161,14 @@ class EditorProvider extends Component {
 			>
 				{ children }
 				<ReusableBlocksButtons />
+				<ConvertToGroupButtons />
 			</BlockEditorProvider>
 		);
 	}
 }
 
 export default compose( [
+	withRegistryProvider,
 	withSelect( ( select ) => {
 		const {
 			__unstableIsEditorReady: isEditorReady,