+++ packages/npm-package-json-lint-config/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 2.0.0 (2019-06-12)
+
+### Braking Change
+
+- Added `type` and `react-native` to the order of preferred properties. 
+
 ## 1.2.0 (2019-03-06)
 
 ### Internal